BUG FIX: PubmedSummary (and PaperAdapter) don't use pubmed info for paper flds
Problem: we were using Geo to grab paper fields, like title. Before, we
had a problem like this where we were grabbing the authors info from GEO--but
the patch just corrected for authors.
NOW: we are grabbing all of the paper fields from pubmed and not GEO!

NOTE: the problem was deeper--PubmedSummary was grabbing scrapping it's
title from the GEO linked query, and not the pubmed query (seriously!).
NOW pubmedsummary comes from pubmed (as it should!)