Admin-only: creating a new level of users based on the is_staff field
We want to create two types of registered accounts, admin and non-admin.

All pages on the dc are restricted to admin-only accounts except for:
1. the homepage (and it's supporting views)
2. the paper submission page
3. the dcstats page

NOTE: in the past, we were using the login_required decorator to see if the
user was authenticated.  NOW we have a admin_only decorator, which not only
checks if the user is authenticated, but also if the user.is_staff.  admin_only
redirects to the appropriate page, i.e. login or restricted page.

To create an admin account, the user's is_staff must be set to True/1