adding a nav bar for administrators
i wanted to integrate all of the pages that the admins/students might need
onto the page itself, so they won't have to remember urls.

NOTE: to do this, i had to hide some of the links, e.g. factors, platforms
under a rollover menu.  this rollover menu was VERY hacked together...it
works, but it's very ugly.