Paper Collection: in Papers and Datasets, all fields are optional!
Fields like paper.gseid and pmid use to be required and unique--this was
too strict of a constraint

So now everything is optional--you can in fact create empty entries

NOTE: the only constraint left- datasets must be associated with papers!

Also: added generic id to both of these classes--just in case they use
a ecbi id or something.