Factors View (tab): Part 3- When num cols > maxCols*2, table must break out!
BUG: when there were lots of columns, the browser was trying to squish them
all in the area defined for the table.

Soln: we define as the maxCols*2 to be the most that could/should fit in the
original area; IF the number of columns exceeds this, we scale the table width
to accomodate them by setting the width to a percentage of maxCols*2.