MAJOR: ncbiAdapter and classifiers- the adapter to interface with ncbi (geo, sra, pubmed) and classifiers to help auto-annotate samples!
This is a major commit--with this commit we achieve a major milestone:
The auto-importing and annotating of ALL chip-seq samples from ncbi resources
(i.e. geo)

The classifiers are PRE-MADE (by the admin--len), and then plugged into the
classifier module (copying in the classifier for each class and the text
file template which to rip).